+++ src/com/reason/lang/QNameFinder.java
@@ -6,7 +6,7 @@ import org.jetbrains.annotations.NotNull;
 import java.util.EnumSet;
 import java.util.Set;
 
-public interface ModulePathFinder {
+public interface QNameFinder {
 
     enum Includes {
         containingFile,