+++ src/com/reason/lang/ocaml/OclQNameFinder.java
@@ -14,8 +14,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static com.reason.lang.ModulePathFinder.Includes.containingFile;
-import static com.reason.lang.ModulePathFinder.Includes.includedModules;
+import static com.reason.lang.QNameFinder.Includes.containingFile;
+import static com.reason.lang.QNameFinder.Includes.includedModules;
 
 public class OclQNameFinder extends BaseQNameFinder {